requirements.txt
SpeechRecognition
numpy
matplotlib
Pillow
easygui
rapidfuzz